requirements.txt
fastapi
uvicorn
librosa
transformers
resemblyzer
spectralcluster
pyannote.core
pyannote.metrics
python-multipart
soundfile
torch
numpy
matplotlib
